Dialogue modifications/cleanup
