+++ Common/VanillaOverrides/VanillaItemOverrides.cs
@@ -69,7 +69,6 @@ namespace OvermorrowMod.Common.VanillaOverrides
 
         public override void ModifyWeaponDamage(Item item, Player player, ref StatModifier damage)
         {
-            //if (player.CheckArmorEquipped(ItemID.CowboyHat)) Main.NewText("a" + player.armor[0].ToString());
         }
 
         public override void ModifyWeaponCrit(Item item, Player player, ref float crit)
@@ -85,7 +84,6 @@ namespace OvermorrowMod.Common.VanillaOverrides
 
         public override void ModifyWeaponKnockback(Item item, Player player, ref StatModifier knockback)
         {
-            //Main.NewText("i" + player.armor[2].ToString());
         }
 
         public override void ModifyTooltips(Item item, List<TooltipLine> tooltips)
