+++ Content/Items/Weapons/Ranged/Farlander.cs
@@ -60,7 +60,6 @@ namespace OvermorrowMod.Content.Items.Weapons.Ranged
 
         public override void Kill(int timeLeft)
         {
-            //Main.NewText("death");
             AICounter = 0;
         }
 
@@ -139,8 +138,6 @@ namespace OvermorrowMod.Content.Items.Weapons.Ranged
             float chargeProgress = Utils.Clamp(gunPlayer.FarlanderCharge / 120f, 0, 1);
             float accuracy = MathHelper.Lerp(12, 0, chargeProgress);
 
-            //Main.NewText("accuracy: " + accuracy + " from " + chargeProgress + " -> " + gunPlayer.FarlanderCharge);
-
             Vector2 rotatedVelocity = velocity.RotatedByRandom(MathHelper.ToRadians(accuracy));
             int chargeDamage = (int)(chargeProgress == 1 ? damage * 1.5f : damage);
 
@@ -158,7 +155,6 @@ namespace OvermorrowMod.Content.Items.Weapons.Ranged
         {
             GunPlayer gunPlayer = player.GetModPlayer<GunPlayer>();
             float chargeProgress = Utils.Clamp(gunPlayer.FarlanderCharge / 120f, 0, 1);
-            //Main.NewText(chargeProgress);
 
             if (ShotsFired < MaxShots) player.scope = true;
         }
