+++ Quests/BaseQuest.cs
@@ -100,7 +100,6 @@ namespace OvermorrowMod.Quests
             {
                 ResetEffects(player);
                 GiveRewards(player);
-                //Main.NewText("COMPLETED QUEST: " + QuestName, Color.Yellow);
             }
             Quests.State.CompleteQuest(modPlayer, this);
         }
